Added Dtype Support

+++ src/python/oscillate/model/layers/ffn.py
@@ -1,12 +1,13 @@
+import torch
 import torch.nn as nn
 
 
 class FeedForwardNetwork(nn.Module):
-    def __init__(self, emb_size, ff_size):
+    def __init__(self, emb_size, ff_size, dtype=torch.float32):
         super(FeedForwardNetwork, self).__init__()
-        self.layer1 = nn.Linear(emb_size, ff_size)
+        self.layer1 = nn.Linear(emb_size, ff_size, dtype=dtype)
         self.relu = nn.ReLU()
-        self.layer2 = nn.Linear(ff_size, emb_size)
+        self.layer2 = nn.Linear(ff_size, emb_size, dtype=dtype)
 
     def forward(self, x):
         out = self.layer1(x)
