+++ src/python/oscillate/model/model/model.py
@@ -7,12 +7,12 @@ from oscillate.model.model.encoder import Encoder
 
 class TTAModel(nn.Module):
 
-	def __init__(self, encoder: Encoder, decoder: Decoder, *args, decoder_vocab_size=1024, **kwargs):
+	def __init__(self, encoder: Encoder, decoder: Decoder, *args, decoder_vocab_size=1024, dtype=torch.float32, **kwargs):
 		super().__init__(*args, **kwargs)
 		self.encoder = encoder
 		self.decoder = decoder
-		self.enc_reshape = nn.Linear(self.encoder.emb_size, self.decoder.emb_size)
-		self.dec_reshape = nn.Linear(self.decoder.emb_size, decoder_vocab_size*self.decoder.emb_size)
+		self.enc_reshape = nn.Linear(self.encoder.emb_size, self.decoder.emb_size, dtype=dtype)
+		self.dec_reshape = nn.Linear(self.decoder.emb_size, decoder_vocab_size*self.decoder.emb_size, dtype=dtype)
 		self.softmax = nn.Softmax(-1)
 		self.decoder_vocab_size = decoder_vocab_size
